requirements.txt
streamlit==1.35.0
langchain-community==0.2.4
faiss-cpu==1.8.0
sentence-transformers==3.0.1
requests==2.32.3
Pillow==10.1.0